Switch over to teamnsrg/chromedp and udpate dependencies
This commit has several changes. We switched over to teamnsrg/chromedp
so that we could add in ModifyCmdFunc and use it to make changes to
the command that will run the browser. We used this to make sure
each browser spawns with it's tmp directory as it's working directory.

This commit also adds a test for closing JavaScript dialogs correctly.